Change switch to laser mode, add video
Toggling the switch now toggles between laser mode and arrow mode
The difference is the projectiles the central turtle shoots: either an arrow (which leaves no trace) or a fast, red laser (you only see the trace, not the "front" of the laser.)

I store the current status of the mode and just toggle it when I get a serial message "Toggle mode" which is an example of storing some state data on the python side. I could also have the arduino say the mode each time with every serial message.... so I thought it'd be an interesting example.

I'm also uploading a powerpoint with a screencap of what's happening on my screen since you won't be able to emulate without your own arduino + breadboard setup.