Add Bluetooth sketch so ESP gives raw data as well
ESP now sends over bluetooth both raw distance data (very approximately in cm) and its instantaneous assessment of if it's available or not in that moment.

The ESP is NOT SMART. Its understanding of availability is based on its most recent sonar reading only. It doesn't consider anything previous. (so it's potentially susceptible to noise, especially with battery variability)